Implement singularity and simplify trimming
All rules now run in their own singularity container.

The trimming of the fastq files was simplified. Before, the contaminants
list from FastQC was used to inform the adapter trimming by Cutadapt.
Afterwards, sickle-trim was used in single end mode to trim low quality
bases. Finally, the fastq reads that went out of sync due to the single
ended trimming were synced back together.

However, in practice, the Illumina adapters which were used during
sequencing are known, so there is no need to parse the output from
FastQC to determine the adapter content. Further more, in practice
the amount of adapter detected by FastQC (~10% towards the end of the
read) was not enough to trigger the contaminants report so that in
practice no adapters were removed at all.

In the current version, Cutadapt is executed to trim the default
Illumina adapters. It also trims low quality bases (Phred < 20) which
matches the default setting in Sickle-trim. Note that the algorithm
that is used by Cutadapt is different from what is used by Sickle-trim.

See
https://cutadapt.readthedocs.io/en/stable/algorithms.html#quality-trimming-algorithm
for details on how quality trimming is implemented in Cutadapt