Fix error in fastqc-processed rule
FastQC was ran on the raw fastq files twice, instead of on both the raw
and processed fastq files. This commit fixes this issue, and add new
tests which use low quality fastq files to verify that the trimming by
cutadapt was successful.

The test data was taken from https://github.com/biowdl/QC