Increase minimum intron size to 50
Any missing sequence smaller than the minimum intron size will be called
as a deletion, rather than a 'free' reference skip due to a presumed
intron.

The reason for this change were two large deletions in the ASXL1 gene
that were missed with the default settings.

(Samples 2-021 and 3-024 from the HAMLET publication).