Manually duplicate consequence for MT-ATP6
This is to mimic the case where a single variant in a gene overlaps
multiple transcripts of interest. This does not occur in the pipeline
test data, so this modified output file is added to be able to test this
use case.

If changes are made to the output format of HAMLET, this file will need
to be updated as well.

It was generated by taking the SRR8615409.summary.json output of the
test-hamlet-chrM test case, and manually duplicating the first
transcript_consequence for MT-ATP6.