No longer reorder the bam file
Previously, we need to reorder the BAM file to match the order of the
contigs in the reference. Note that this is different from sorting by
coordinate, which only sorts reads within a contig.

However, STAR already outputs the contigs in the same order as the
reference, so this step is no longer needed.