Add xz compressed reference files
The trick here is to N-mask all chromosomes, except for the genes of
interest (see hamlet-ref.bed). If you compress the resulting file with
xz, it becomes extremely small. When running the functional tests, we
first unpack the .xz files, which is very fast. The biggest advantage of
this is that even with small test data, all chromosomal positions are
maintained.