Use MT-CO2 to test the expression module
MT-CO2 was used both as gene of interest, and as sole housekeeping gene,
which means that the relative expression must be 1.0 for every
strandedness.

MT-CO2 was chosen specifically because it does not overlap any other
transcripts, which would have been taken into account by STAR, giving a
lower count and throwing off the relative expression.