Speed up VarDict by sorting call_regions
When running VarDict with multiple threads, the analysis is parallelized
by dividing the bed records over the threads. However, not all bed
records have the same size, and some of the largest records are towards
the end of the file (chrX). This leads to a long runtime for the job as
towards the end, only one of the threads is still active.

By sorting the bed file and putting the largest records first we can
ensure better utilization of the available threads, and hence faster run
times.

The code of VarDict which writes the final VCF ensures that the VCF file
is properly sorted.