Add MECOM exons to use to quantify gene expression
MECOM-206-e1 is the first exon of EVI1, which is not shared by the
MDS1-EVI1 fusion transcript. It has been reported that only high
expression of EVI1, and not MDS1-EVI1 is correlated with poor prognosis
in AML.

MECOM-220-e15-e16 is the region that overlaps exon 15 and 16 of the MANE
Select for MECOM, which is the MDS1-EVI1 fusion transcript.