Run integration test sequentially
By running the entire pipeline first, and by running all pipelines
sequentially, we ensure that all singularity images that are needed for
HAMLET are in the cache after running the first job.

This will allow us to only cache one set of singularity images, while
ensuring that all needed images are available for every module.