Update VEP to version 115
This comes with some changes in how transcripts are annotated from a gtf
file, so we had to remove the transcript version from all test cases.

Briefly, VEP never used the transcript version that was defined in the
gtf file, but always defaulted to 1. Now, VEP does not include the
transcript version in the HGVSc description at all, see this bug report
with VEP for more information:
https://github.com/Ensembl/ensembl-vep/issues/1933